requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic-settings==2.4.0
azure-storage-blob==12.22.0
azure-identity==1.17.1
